Move CLI command implementations to commands directory
